fix: gitignore should ignore .toml/.conf files
As the application requires a config.toml or config.conf file it makes
sense to ignore these types of files rather than having them show up
in git status output.

While the files are technically in the toml format, we use the .conf
extension in our documentation so ignore both.