feat: add custom time/date format field for elasticsearch_query
This adds the ability to specify a custom time/date format to
elasticsearch queries. Elasticsearch has a large number of built in
date/time formats, however users are also able to add in a custom
date/time format as well. If this is the case, then the query will
always fail as it will be unable to parse the date/time.

This field is generally not required if a user is using a
standard/built-in type.