fix: trim whitespace on ethtool input plugin
The upstream vmxnet3 driver declares the network driver stats
descriptions with leading and trailing whitespace. This was done to help
with the ethtool output on the CLI. The ethtool plugin uses a library
that polls the driver's stats directly and therefore this ugly
whitespace will show up in Telegraf output as well.