fix: json_v2 parser timestamp setting
The json_v2 parser does a carterisan product during parsing. As part of
this product, it will attempt to merge metric fields and tags. However,
timestamps are not part of this merge process. In the current state,
the custom timestamp will always get overriden if the timestamp field is
the last value in the JSON.

Adds a test and does modify one of the tests, which produces the exact
same metrics, but in a slightly different order.

Resolves: #10160