fix: do not require networking during tests
There are a couple of tests that get run which reach out to external
network devices. Unit tests should not depend on functional networking
or be reaching out to random servers. The DNS server is somewhat
understandable, but the SNMP server is an unknown system that is not
owned by the development team.

Fixes: #9894