chore: update mqtt documentation
There is some confusion reguarding the server's string as well as the
need to use the keep_alive param with later versions of mosquitto. This
tries to make the README read a bit more clear and not need two
different sections, which detail the various parameters.

Finally, it gets the README and the configuration in sync with each
other.

Fixes: #10180