fix: ensure http body is empty
Even if the http body was empty, the current method will create an
io.ReadCloser and send that to the http request. This was causing a 502
error from a service. Ensuring that the empty body is actually empty
prevents the server side error message.

Add log fixture to tests to aid in debugging.

Fixes: #10386