feat: add consecutive_name_prefix_limit to Aggregator, Input, and Output plugins, such that the prefix is not appended if a metric name already begins with N occurrences of the prefix, where N is the defined limit.
If the limit is unset, the prefix will always be applied.

Using the prefix `foo_` and a `consecutive_name_prefix_limit` of 2, a metric
with the initial name `foo_bar` would translate into `foo_foo_bar`, i.e.
the prefix would be applied. A metric with the initial name `foo_foo_bar`
would remain `foo_foo_bar`, i.e. the prefix would not be applied.

Signed-off-by: Conor Evans <coevans@tcd.ie>