feat: always disable cgo support (static builds)
This ensures that cgo is always disabled with builds using
the Makefile. Telegraf does not support importing C libraries
or depending on any of them in the first place. This is to
ensure cross-platform and cross-OS support.

Our official builds for Windows, macOS, and BSD are
cross-compiled. In these cases, cgo is already disabled. This
primarily affects the linux builds. This will enable us to deliver
the much requested arm64 docker images for Alpine.

In #8278, a check for the glibc version was added after a
user noticed the binary was depending on newer versions of
glibc. This removes those checks as static builds will not need
these checks.

Finally, this removes the no longer necessary static specific
build.