feat: always disable cgo support (static builds)
This ensures that cgo is always disabled with builds using
the Makefile. Telegraf does not support importing C libraries
or depending on any of them in the first place. This is to
ensure cross-platform and cross-OS support.

Because Telegraf does not need to depend on dynamic libraries,
we can ensure Telegraf builds statically.

MacOS and Windows do not have official static build support and
by default builds on those systems are already as static as they
can be. This primarily affects the Linux/*BSD builds.

Finally, this removes the no longer necessary static specific
build.