tests: reduce unit test time
The unit tests take upwards of a minute on a very fast system and over
four minutes in CI. There are a few changes that were made:

* Some tests are just long running tests, more integration style
tests or have hard-coded sleeps that cannot be changed or fixed. These
types of tests have been skipped.
* Some sleeps and/or timeouts have been reduced down from multiple
seconds to milliseconds. This way the tests have the opportunity to
end faster.
* http error code 103 in the go library causes at least a 2 second
delay in unit tests. These have also been skipped in short mode.
* increase windows size to 4 cores, like other systems
* remove leofs integration tests