fix: remove signed macOS dotfile artifacts
When the macOS enviornment is attached to the package-consolidate
CircleCI job, there are two new files. These dot-underscore files are
hidden files for macOS that store metadata about files. These files are
not required and not useful for our release process. In fact, they cause
issues when they show up.

This will remove any of these dot-underscore files before artifacts are
uploaded to CircleCI.