fix: use systemd detection in RPM install scripts
Instead of having an ever growing list of distros and releases, this
collapses the logic to switch based on whether systemd is detected or
not.

This also removes the sysv logic as all supported Fedora and RHEL
distros are only systemd at this point. This does leave the check for
systemd in place to ensure installs in containers work without errors.

Fixes: #4507