fix: use now as ecs timestamp
The current ecs timestamp option will use the timestamp when an image
was last pulled down. Because the tags are not changing for a specific
ecs container this means that the metric never actually updates.

By changing the timestamp to be the current time, which is more standard
for metrics, users will get updated information for their fields.

Fixes: #8375