fix: sql allow specific unsigned settings
Allow the user to specify a specific unsigned value to use for
converstion rather than assume the integer value + unsigned. This is
done via a new option.

Additionally, this actually adds tests for the uint64 and float64
datatypes across clickhouse, mariadb, and postgres. Because postgres
does not have support for unsigned values, the number is treated as a
bigint.

Fixes: #10671