chore: remove parallelism and spliting of CI tests
The CI system was setup to split out and run the unit-tests across 4
different nodes. When parallelism is enabled it will create and build 4
different CI systems, which do not all happen at the same. Given that CI
tests take a matter of seconds this change does not save us time and
instead costs us more in credits than it is worth and takes longer
overall.