chore: skip longer/integration tests on -short mode
This is a collection of longer running tests that do not have a clean or
simple solution to shorten their length or reduce their system
utilization. Some cases are due to large memory usage or due to
hard-coded seconds long sleeps.

The leofs tests are integration tests that are actually compiling
in-line code as a go binary.

The vsphere tests make a very long call to create a simulator.

In each case, these can be skipped when running with short.