fix: redundant zfs pool tag
If a ZFS pool has multiple datasets it will show up in the ZFS pool list
that Telegraf maintains. When trying to produce the ZFS pools tag this
will result in a pool name showing up more than once. This creates a
unique set of pool names for the tag.

Fixes: #10837