fix: default value for logfile rotation interval
The unit of 'd' is not a valid time.Duration unit. Due to the more
strict parsing of durations, this is now correctly failing. Previously
using a value of '1d' would fail silently. If a user wishes to use a
value of days, they need to convert to hours and use '24h' for example
of one day.

Fixes: #10882