chore: add logging to graylog tests
The TCP graylog tests are one of the most flaky tests we have. The
current scenario attempts to send two messages, then close the
connection, fail to send a third, and then correctly send a fourth.

I was unable to reproduce the failure after a couple dozen times in CI
and after running the tests >10,000 times locally. For now, this adds
some logging to stdout so we can see what might be going on at the time
of failure.