fix: update vsphere info message to debug (#10903)
This message is very verbose and causes a lot of extra messages to get
printed out as-is. Having it can be helpful in debugging when a metric
does not show up, but it is not useful for general usage. This changes
the message to only show up with debug enabled.

Fixes: #10350
(cherry picked from commit b169b015680e68d265d14a2771ba240d702037c5)