fix: add tracefs to ignore, reduce log level
The tracefs filesystem is showing up more and is not likely something a
user wishes to have enabled by default. This adds it to the ignore in
the config by default.

This also reduces the error message level from error to debug. This
was previously a debug message, but was bumped all the way to an error.
This message does not prevent any messages from getting read by Telegraf
and not an error that halts Telegraf. Users may still wish to address
the messages so reducing to a debug is better than hiding them.

Fixes: #10897