fix: remove overly verbose info message from cisco mdt
When messages are received from Cisco MDT devices the message is scanned
for the `keys` and `content` fields. If these are missing then there is
no parsable data for Telegraf. If lots of messages are coming in, this
message can spam the Telegraf log with lots of uncessary messages on the
order of hundreds per second.