fix: duplicate influxdb listener writes
The influxdb_listener input plugin, when using the upstream parser had
bad logic that would write first using the new upstream parser and then
attempt to parse the data again with the internal parser right after.
This was leading to EOF errors on the internal parser because the data
had already been read.

Fixes: #10900