chore: remove go generate from CI tests
go generate command is unfortunately a single-threaded, one at a time
command. There are no flags to change this. As a result it takes a long
time in our project due to the large number of plugins.

We run this command in CI twice, once during tests and again during
package build. The tests do not need this to be run and this removes it
to save some time on the overall CI process.