fix: re-init azure monitor http client on context deadline error
The context deadline exceeded error most commonly indicates that a
network related issues has occured preventing the client from getting to
the server. Usually this means that there is something the user needs to
address. However, there appears to be a case where a network issue (e.g.
a down proxy) can occur and telegraf never recovers from the down
service.

This checks for the context deadline error and will re-create the HTTP
client if it is hit.

Fixes: #10950