fix: have telegraf service wait for network up
The current behavior for the systemd service file is to wait for the
network.target to activate. This does not mean that the network is
actually up or that a network device has an IP address. The
network-online.target instead waits for a functional network to be up.

A user recently saw issues with this when trying to use a remote config.
On boot, the network.target was hit, but the wireless network was not
ready yet, and telegraf failed to obtain the remote configuration and
never started.