fix: do not error when closing statsd network connection
Errors from the UDPListen function were previously not checked. In a
recent minor release we started checking them, which introduced more
visible errors to users. However, when statsd is run with the `--test`
option or when a user kills Telegraf with a ctrl-c, an error now shows
up as the read is still trying, after a close has occured.

The ReadFromUDP call blocks untill something is received, so I don't
believe there is anything else for us to check to prevent.

Fixes: #9867