fix: remove any content type from prometheus accept header
In #6745, the prometheus accept header was changed to accept any media
type. However, our prometheus plugin only accepts text. With the release
of newer versions of prometheus, the OpenMetrics type is now available
and could potentially be setup as the exclusive response type. As this
new content type is not supported, Telegraf should not accept it.

The original issue, #6523, was filed around getting a 406. The issue had
comments from the rabbit-mq maintainers who made changes to their code
to be less regid resolving the issue. The change to telegraf was made
afterwards anyway.

fixes: #10248