test: switch internet speed to enable file download
The file download test when enabled is significantly faster in terms of
total test time (45 sec vs 4 sec). In CI when running all these tests
in parallel we could hit limits or timeouts, so we want to limit our
need for external networking.

Running with the previous setting, while running other integration tests
can lead to timeouts and other networking hickups, which cause this test
to fail.

fixes: #11123