test: add coveralls coverage to master
On amd64, linux unit test runs, code coverage will be collected and
uploaded to the coveralls coverage service on runs with the master
branch. This does not include forks or PRs from community members.

Currently, the coveralls service requires a token to be used to upload
all coverage reports. However, there is no safe/secure way with circle
ci to hide this with our current processes that we have in place.

This also removes the coverage for the parsers/influx/machine.go file as
it contains many "//line" comments which currently break reports. See
golang/go#35781 for more details.

fixes: #10733