test: update wait for statement for postgres
The postgres docker image will start up and immediately shutdown the
service will the docker entrypoint starts configuration. This results in
the expected ready to accept connections log message to show up twice.
This generally happens within a tenth of a second, but in CI this can
get hit and the test will fail as the database is not yet up.

fixes: #11292