docs: explain directly connecting to mongo node
By default, the mongo-driver library will attempt to use the primary
node of a cluster. For monitoring with Telegraf, users may specify a
secondary node that they want to monitor. In these cases, the user will
need to specify a direct connection to that node.

The library has a number of options for how it connects, which include
primary, primary preferred, secondary, secondary preferred, and closest.
None of these are options to ensure that the specified node is used.

fixes: #11275
fixes: #9555