models/running_output.go: Add systemd Watchdog support
Add support for `Type=watchdog` in the systemd unit.

Change `WatchdogSec=` to the lowest output interval * 2.

If an output is not notifying systemd within the timeframe of WatchdogSec,
it will be stopped. This will catch the case when there is no output that is
working properly and make it easier to monitor the service from systemd.

As there is a debate if this is how telegraf should work or not, Type in the
systemd unit is not changed from its current value.