test: add nightly package testing to amd64/linux
This adds nightly testing of the generated amd64/linux RPM and DEB
packages. This uses LXD in a VM on CircleCI to launch a number of
different distros. The official InfluxData repo is added per the
official instructions, then the new RPM/DEB is installed. This
method verifies that the upgrade works as expected and the service
continues to run.

The package_lxd_test can be used to also locally when doing development
or changes to the DEB and RPM scripts. LXD is required for it to
correctly works. This is done as LXD let's us use system containers to
verify the systemd connection with the services versus docker, which
only lets us sort of test that the install works, but not the end-to-end
service.

fixes: #10467