Add support for retrieving SSD endurance information from Smart Data where it's present (related to #8701)
There isn't a standardised attribute ID (or name) for SSD endurance information with SMART data.

Some vendors use ID 202, others use 177 (and others use others).

The two following entries indicate the remaining lifetime of two different SSDs

    202 Percent_Lifetime_Remain P---CK   094   094   000    -    6
    177 Wear_Leveling_Count     ------   100   100   050    -    432

Because those device manufacturers that *don't* use `202` might use `202` for other information, we cannot safely rely on the ID to identify the information.

This commit introduces `deviceFieldNames` which is a map of names (such as `Percent_Lifetime_Remain`).

When the SMART data is iterated over in `gatherDisk()`, each entry's name is checked against `deviceFieldNames` andd a field created where a match is made