fix(inputs/kube_inventory): send file location to enable token auto-refresh
Currently, Telegraf will read a token file and provide that as a string
to the K8s client. This will prevent the K8s client from attempting to
auto-refresh the token. In newer versions of K8s, tokens will expire
every hour as a security measure.

This instead will pass the file location rather than having telegraf
read it directly. Becuase of this new K8s behavior, this also deprecates
and warns the user when using a hard-coded string token.

fixes: #11267