fix(agent): add flushBatch method
When Telegraf processes a very large amount of data, such that the
batch size is constantly hit, then Telegraf may be sending batches
constantly. At the same time, the flush interval could be hit. During
the sending of a batch, the ticker for the flush interval can go to
zero, and as a result, the batch will produce a warning about not
sending within the interval time.

Sending batches should not send messages to the user that the interval
expired as they are sepearte paths.

Previously, #8953 attempted to reset the ticker during the sending of a
batch. This introduced an issue where the buffer was never fully
flushed.

fixes: #8941