docs: explain PowerShell 5 encoding
PowerShell 5 is the default version of PowerShell in Windows 10 and
older versions. It does not default to UTF-8 encoding, which can cause
issues with Telegraf trying to parse the file.

Rather than adding lots of weird logic to Telegraf, document the
requirement to specify the encoding when using PowerShell 5.

Newer versions of PowerShell now use UTF-8 encoding.

Fixes: #11614
Fixes: #4880
Fixes: #6662