fix: initialize parser in init vs every gather
The grok parser builds a custom set of patterns everytime a parser is
initialized. Currently the file plugin will create a new parser at every
single gather cycle for every file. This was leading to an ever growing
increase in memory usage of extra grok parsers that had not yet been
garbage collected.

fixes: #11809