build(rpm): Allow building RPMs cross-OS
The RPM build process uses all of the cross-compiling infrastructure to
build an RPM that contains correct binaries for the target OS, but,
unfortunately the RPM itself would be tagged for the host OS, preventing
installation.

Now we use GOOS for the target OS for the RPM as well, making
cross-building possible.