fix(inputs.zookeeper): add the ability to parse floats as floats
Zookeeper currently sends anything not able to be parsed as an int to
become a string. This means even other numerical data like a float
becomes an string. This add an optional setting to let the user attempt
to parse floats. If that fails then the field still becoms a float.

fixes: #7567