docs(aggregators.derivative): Remove incorrect comment
This comment states that *if* the variable option is used, the name of
the resulting variable is generated differently. However, in practice
the name of the resulting variable will always just use the (specified
or default) suffix, regardless of whether the variable option is set.

It seems this behavior and this comment were both already present when
the plugin was first introduced in PR #3762 (commit 927d34f), probably
a remnant of changes during PR discussion.