docs(serializers.json): specify version of jsonata used
The JSONata spec has rolled a number of times and as such added a number
of other very helpful functions for use. Unfortunately, the library that
telegraf uses has not been updated to support these newer versions. At
first glance there also does not seem to be a replacement library at
this time.

fixes: #12070