test(inputs.mongodb): update integration test logic
These tests were expecting specific log messages, that were flaky. While
looking at these tests I also realized that these should be testing how
servers are handled not with changes, but from the start. Therefore,
this updates the logic a bit to ensure a) on error, we actually error
out quickly and b) on skip, that no errors are returned.