fix(agent): fail on start error for test and fail modes
This aligns the behavior of the --test and --once modes to match that of
the general telegraf operations. Currently, if there is a failure during
a call to an input's start telegraf will stop. This is not the case for
--test and --once modes. In these cases, the error will print out, but
not stop execution. This can lead to crashes due to some variables not
getting correctly initialized or setup in Start.

fixes: #12350