test(inputs.statsd): increase sleep + expiration on test
This test waits for 3 metrics to appear. On Windows, in CircleCI this
test has become flaky and timing out after 10mins waiting for those
metrics to show up. My hunch is that the microsecond sleep and
expiration are too short and small even when the Windows machines tend
to be very slow.

fixes: #12341