chore: do not start service on deb install
Because the default config is no longer valid, telegraf will error out
and the systemd service will be marked as failed. Instead, no longer
start the service on install.

Additionally, on an upgrade, if the service is already running, then
restart the service, but again do not start it if it was not already
started.

fixes: #12397