fix(agent): only set default snmp after reading all configs
Currently the default snmp reader is set after reading every file. This
can lead to unnecessary deprecation messages that do not make sense to
the user. Instead, only set the default, deprecated snmp reader after
all files are read in. This way the user is only notified of the
deprecation if they actually choose to use that value in their config.

fixes: #12454