fix(bond): reset slave stats for each interface
Previously the `status` value has been initialized to 0 only when
starting to parse a whole bond. If any single interface in the bond is
up the `status` is set to 1 for all following interfaces (even if they
are down).

To fix this we reset all collected fields after each interface is
finished. This should also resolve the unlikely case that some
interfaces have different fields than others.