fix(inputs.upsd): always convert to float
There are a number of values that can show up as a float or an int. This
ensures the floats are always floats to avoid having some devices report
as an int and others as a float and conflict in influxdb.

Additionally, if we fail to convert in either the float or string case,
then skip trying to add the field.

fixes: #12501