docs(inputs.postgresql): provide socket connection details
This adds notes about connecting to postgresl using a local unix socket.
While the issue below was hopeful in changing the default behavior at
the very least documenting that you can connect to a local socket via
the connection string is helpful to users.

The default will remain the same however.

fixes: #4872