feat(agent): add /etc/telegraf/telegraf.d to default config locations
This will add /etc/telegraf/telegraf.d to the list of default
configuration locations to load from. Currently, users need to specify
an empty /etc/telegraf/telegraf.conf file and specify they want to
include the configuraiton directory outside of the RPM and DEB.

A common use case is for Docker where users want to upload an entire
directory of configurations and run telegraf. Right now they have to add
a custom launch command, when in reality we should include this
directory by default.

This also means that the --config option can be omitted from the CLI
if a user is only using --config-directory.

fixes: #5656
fixes: #5571