feat(inputs.internet_speed): server ID include and exclude filter
This adds the ability to exclude or include specific speedtest server
IDs. If a user excludes a server ID that server is always excluded from
selection. If a user includes a specific server ID, then we will look
for only that ID.

This does not let a user pick a random server ID. The speedtest-go
library will receive the 10 closet servers. As such, users need to
be careful as they can prevent any server from being picked.

fixes: #11449
fixes: #11625
fixes: #12595