docs(inputs.syslog): Add BSD syslog to readme intro
Whilst support for the BSD syslog protocol (RFC 3164) has recently been added to this plugin, the introduction section still stated that only the newer RFC5424 protocol was supported. This was despite the subsequent sections of the readme explaining how to optionally use the BSD syslog protocol.

Update the introduction section of the readme to clarify that both the RFC5424 or RFC3164 protocols are supported.