fix(agent): reload config on config update
In a recent change we moved to loading the config only once during the
initial load of telegraf. This saves time and is more efficient from
reading files on disk, however, it ignored the fact that we need to
reload the config if someone is watching a config. This reload occurs in
runAgent, where the second reading was always taking place.

This updates the logic, so that we only re-read the configs in runAgent
if we are calling it a second time from the reload loop.

fixes: #12718