fix(agent): Only rotate log on SIGHUP if needed
The current code will rotate a log on a SIGHUP. Whether the log needs to
be rotate or not. There are checks at start up and at each write if a
log needs to be rotated. Meaning that we should not be forcing a rotate
at close, and instead check to verify the conditions are even met first.

fixes: #12709