chore: unify the secretstores config
Because the config that is embedded in the telegraf binary is always
generated using a linux binary during releases, we need the config to be
unified across all platforms (e.g. macOS, Windows, and Linux). This puts
the configs as one sample config, with some additional messages to
clarify what applies to each platform.