chore: use type on powershell/command prompt only
This will use cat in the makefile for all systems and prompts, except
for on Windows where a user is using PowerShell or Command Prompt. That
way existing usage of bash on Windows continues to work as expected.

Using file is only supported on Makefile 4.2.1 and newer. Which is not
available in OS X and older RHEL 7 for example.

fixes: #12573