fix(outputs.stackdriver): group batches by timestamp
In a perfect world, metrics come in to stackdriver grouped by timestamp
and all the timestamps are the same. This means that metrics go out
together and a user never has to worry about out of order metrics.
The world however is not perfect.

In the event that the connection to stackdriver goes down, telegraf will
start to save metrics to a buffer. Once reconnected, telegraf will send
a batch size of data to stackdriver to send. Stackdriver would then
ensure the metrics are sorted, but then break the metrics into
timeseries. In doing so metrics would no longer be in any order and
metrics could send duplicate time series. Meaning two idential metrics,
but with different time stamps.

What the user would see is first a message about duplicate timeseries
and then an error about out of order metrics, where an older metric was
trying to get added.

This ensures that we avoid different timestamps by batching metrics by
time. This way we avoid duplicate timeseries in the first place and
ensure we always send oldest to newest.

Fixes: #12963