fix(inputs.socket_listener): fix loss of connection tracking
Previously if an error was encountered during socket initialization, it would return early from the setupConnection function, and skip adding the connection to the tracking list. This could cause issues later down the line.

This refactors the code so that setupConnection() only returns an error if the connection should be aborted. Otherwise messages should be written to the logger.

fixes #13051