fix(inputs.socket_listener): fix tracking of unix sockets
This fixes an issue where the code would lose track of unix sockets. The remote end of a unix socket does not have a unique address representation, thus multiple entries may overwrite each other in the map. This changes the map to key off the net.Conn object itself, basically using the map as a set.

Fixes #13058