feat(processors.parser): add merge with timestamp option
By default, the merge processor will use the existing timestamp of the
metric. This makes somesense when the user wants to keep the timestamp
of when the metric was originally created. However, there are cases
where if the user is also parsing data at this stage that they may want
to set the timestamp as well.

Fixes: #13102