fix(processors.starlark): do not reject tracking metrics twice
If an error occurs in the starlark processor tracking metrics are
rejected, decrementing their count. However, immediately after the agent
when running processors, if an error occurs in the processor it will
also attempt to drop the metric. This double reject + drop results in a
negative reference count, panicing telegraf.

This did not affect metrics which do not use tracking metrics. It
requires a metric that uses a tracking metric to expose as well as an
error in starlark.

fixes: #13148