chore: update removal and deprecations for v1.30.0
This updates the various locations that v2.0.0 is used as the removal
date for telegraf. As we are not doing a v2.0.0 anymore, we still want
to remove many long standing items. Using the v1.30 release is a
good, round number to make that happen and this gives folks more time.