chore: produce single digest file with sha256sum
This will update the name of the digests file that is produced to ensure
it is uploaded to S3. Additionally it removes the un-needed signatures
and switches the order so that the hash is listed first, then the
filename as is more common. As this is already produced in the
individual digests we are essentially only combining them all into one.