fix(inputs.mqtt_consumer): Release semaphores on errors
The recent changes to the mqtt_consumer to ack messages would not
release semaphores when an error occured during parsing. This resulted
in Telegraf to seemingly to receving metrics once it reached the limit
of the max undelievered messages.

The change ensures that whenever we error or do not have any metrics
during the initial message handling we also decrement the semaphores.

Additionally, this ensures that we only ack messages when persistence is
enabled.

fixes: #13476