feat(http): Add support for connecting over unix-socket
The `inputs.http` and `outputs.http` now supports connecting over
unix-socket. The unix-socket path can be specified via
`unix_socket_path` configuration parameter. When this is set to a
non-emtpy value, we connect over a unix-socket.

Note: Even when connecting over unix-socket, the value for the `urls`
must still begin with "http://". The hostname is not really important
in such case and so it can be set to anything. However, "localhost" is
recommended to avoid any confusion.

Example:
```
[[inputs.http]
    urls = ["http://localhost/data/endpoint"]
    unix_socket_path = "/tmp/foo/server.sock"
```

Fixes #11038