lustre2: cleanup test suite by adding rootdir
Now, the testsuite can write mock procfs files to standard locations
within the tmpdir without needing to specify which files are used.

This is really just a precursor for future commits. We want this plugin
to "just work", without needing to add more fileglob fields to the
config file for each one. Instead, the plugin will search on the
standard paths. We won't remove the old .toml fields for backwards
compatibility, but they will be deprecated and we will avoid adding new
ones. This commit makes it possible to continue using the testsuite in a
similar pattern, but we'll only need to fill in "rootdir", rather than
globs for each new type of file.