fix(inputs.prometheus): fix missing metrics when multiple plugin instances specified
Plugins share SharedInformerFactory so that they reuse single connection to K8S API
servers. If a plugin limits scope to just one namespace it can't share
InformerFactory with another plugin which specifies different namespace.

To handle it correctly create a map of Namespace -> InformerFactory so that
plugins can hook up to the right one.