test(inputs.radius): Remove local test
Instead of using what could be an already busy port, rely on the
integration tests for testing radius. The advantage is that the
integration tests will randomize ports when using test containers to
ensure that a port is not already in use.

This prevents random test failures where a port is already in use.