feat(serializers.nowmetric): Add option for JSONv2 format
The nowmetric serializer by default produces metrics in the Operational
Intelligence format. This adds a format object, which can be set to
jsonv2 to produce metrics compatable with the JSONv2 endpoint.

The primary difference is in the JSON format. The JSONv2 endpoint is
based around an object, while the previous OI format is based around a
list of metrics.