feat(inputs.procstat): new option to select which metrics should return
Add a new option, metrics_include, to the procstat input plugin to
select which metrics should be returned for each process.

This commit prepares the input plugin to optionally gather more data
without modifying the default set of values returned.
It also avoids the cost of gathering the data and then discarding it
with a fielddrop.