feat(inputs.procstat): add tcp related metrics
For a given process, return the number of connections in each state:
SYN_SENT, SYN_RECV, ESTABLISHED, etc.

Also added a new measurement, procstat_tcp, that returns the endpoints
where the process is connected to and where it is listening.
If it is listening in "[::]" or "0.0.0.0", resolve with the local IPs.