feat(input.intel_pmt): Add pci_bdf tag to uniquely identify GPUs and other peripherals
In the current intel_pmt plugin, the `numa_node` of the intel_pmt device is used as a way to uniquely identify which CPU socket the sample belongs to. However, with GPUs and other future peripherals, an additional tag is needed to uniquely identify samples.

Thus add the parent of intel_pmt's device PCI Bus:Device.Function (BDF) as the tag `pci_bdf`. We find the PCI BDF by traversing up from `telemDeviceSymlink` and taking it's basename. Equivalent example in bash:

```
$ basename $(realpath /sys/class/intel_pmt/telem10/device/..)
0000:e7:03.1
```

All intel_pmt parent devices appear to be PCI devices, so we shouldn't need to special-case this tag between CPUs and other peripherals.

Sample including pci_bdf:

```
> intel_pmt,datatype_idref=ttemperature,guid=0x9956f43f,host=node001,numa_node=0,pci_bdf=0000:e7:03.1,sample_group=TEMPERATURE[4],sample_name=CORE_TEMP value=36.4140625 1695671600000000000
```